+++ Write-Log.ps1
@@ -1,5 +1,6 @@
 function Write-Log {
     param([string]$logstring)
     Write-Debug -Message "Append ""$logstring"" to log file: ""$logfile"""
-    Add-Content $logfile -Value $logstring
+    #35
+    $logstring | Out-File $Logfile -Append -Encoding utf8
 }